requirements.txt
asgiref==3.8.1
bleach==6.2.0
crispy-bootstrap4==2024.10
Django==5.1.2
django-admin-adminlte==1.0.5
django-crispy-forms==2.3
django-summernote==0.8.20.0
gunicorn==21.2.0
packaging==24.1
pillow==11.0.0
python-dotenv==1.0.1
sqlparse==0.5.1
str2bool==1.1
webencodings==0.5.1
whitenoise==6.6.0
